refactor(state/project): extract §4 inline business logic to helpers.ts (ADR 0005 §4)
Move rewriteFolderTreePaths, basename, renameFolderInTree, and moveFolderInTree
out of project-slice.ts into a co-located helpers.ts. The slice now delegates
all folder-tree mutation logic to these pure functions; no behaviour changes.
